backend/requirements.txt
# Backend dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
pydantic>=2.5.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
httpx>=0.26.0

# Telegram Bot
aiogram>=3.3.0

# Utils
aiofiles>=23.2.1
